Add minor linting fixes and moved ListItems in components and CardContainer into containers
