+++ src/containers/Form/index.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react'
-import ListItem from '../ListItem';
+import ListItem from '../../components/ListItem';
 import { addNewNote } from '../../thunks/addNewNote';
 import { connect } from 'react-redux';
 import { updateNote } from '../../actions/index';
