+++ src/index.scss
@@ -1,6 +1,6 @@
 @import'./containers/App/normalize';
 @import './containers/App/App';
 @import './containers/NoteCard/NotesCard';
-@import './components/CardContainer/CardContainer';
-@import './containers/ListItem/ListItem';
+@import './containers/CardContainer/CardContainer';
+@import './components/ListItem/ListItem';
 @import './containers/Form/Form';
